refactor. step5 - basicGame과 컨트롤러 분리
